Move `State` class to dedicated file
