+++ src/collect/collect.ts
@@ -2,7 +2,8 @@ import { CallbackHandler } from "../callback/callback";
 import { AbstractCollector, CompleteInvoice } from "../collectors/abstractCollector";
 import { CollectorLoader } from "../collectors/collectorLoader";
 import { AuthenticationError, DesynchronizationError, LoggableError, MaintenanceError, NoInvoiceFoundError } from "../error";
-import { IcCredential, State } from "../model/credential";
+import { IcCredential } from "../model/credential";
+import { State } from "../model/state";
 import { Customer } from "../model/customer";
 import { User } from "../model/user";
 import { Location } from "../proxy/abstractProxy";
