+++ src/collectors/abstractCollector.ts
@@ -2,7 +2,7 @@ import axios from 'axios';
 import { Location } from '../proxy/abstractProxy';
 import { Secret } from '../secret_manager/abstractSecretManager';
 import { TwofaPromise } from '../collect/twofaPromise';
-import { State } from '../model/credential';
+import { State } from '../model/state';
 
 export enum CollectorState {
     DEVELOPMENT = 'development',
