+++ src/collectors/apiCollector.ts
@@ -5,7 +5,7 @@ import { mimetypeFromBase64 } from '../utils';
 import { Location } from "../proxy/abstractProxy";
 import { Secret } from "../secret_manager/abstractSecretManager";
 import { TwofaPromise } from "../collect/twofaPromise";
-import { State } from "../model/credential";
+import { State } from "../model/state";
 import * as utils from '../utils';
